Add basic code to Makefile and server.c
The makefile should be fine as it is. However we have some useless
files. If server.c and client.c have their own main function, they
should not have a header files and main.c should not be included
in our files. We are also missing netCode.h and netCode.c.