Add getSocket function, begin initConnection function code
I added the getSocket function to server.c for now just for testing
purposes. In the future getSocket should be a general function
in netCode.h that the client and server should be able to use. I
also began the code for the initConnection function that will handle
the handshake with the client and server.